PM-360: fix action renaming
