+++ src/integrations/baseIntegration.js
@@ -121,6 +121,13 @@ class BaseIntegration {
     this.network = undefined
     this.account = undefined
     this.walletEnabled = false
+
+    this.runProviderUpdate(this, {
+      account: undefined,
+      balance: undefined,
+      network: undefined,
+      available: false,
+    })
   }
 }
